Add files via upload
This project applies data science and machine learning techniques to predict the survival of passengers aboard the Titanic using historical passenger data. The dataset used for this analysis is the famous Titanic dataset from Kaggle, containing details about passengers such as age, gender, class, and more.

Project Overview
The main objective of this project is to build a predictive model that can estimate the likelihood of a passenger surviving the Titanic disaster based on various features like age, gender, ticket class, and more.